+++ providers.go
@@ -57,7 +57,7 @@ func ProvideConfig(logger *slog.Logger) (*Config, error) {
 	}
 	// Override from CLI flag
 	if cli.NoCleanup {
-		config.RunInShell.NoCleanup = true
+		config.RunShellCommand.NoCleanup = true
 	}
 	logger.Info("configuration loaded")
 	return config, nil